+++ MicroFlip/exchange_wex.h
@@ -42,13 +42,13 @@ class Exchange_wex : public Exchange
     QString getRequestErrorMessage(QJsonObject *object);
 
   public slots:
-    void receiveUpdateMarketTicker(QString pair, QObject *sender, int SenderID);
-    void receiveUpdateMarketDepth (QString pair, QObject *sender, int SenderID) ;
-    void receiveUpdateMarketTrades(QString pair, QObject *sender, int SenderID);
+    void receiveUpdateMarketTicker(QString pair,    QObject *sender, int SenderID);
+    void receiveUpdateMarketDepth (QString pair,    QObject *sender, int SenderID) ;
+    void receiveUpdateMarketTrades(QString pair,    QObject *sender, int SenderID);
     void receiveUpdateBalances    (QObject *sender, int SenderID);
-    void receiveCreateOrder       (QString pair, int type, double rate, double amount, QObject *sender, int SenderID);
+    void receiveCreateOrder       (QString pair,    int type,        double rate, double amount, QObject *sender, int SenderID);
     void receiveCancelOrder       (quint64 orderID, QObject *sender, int SenderID);
-    void receiveUpdateActiveOrders(QString pair, QObject *sender, int SenderID);
+    void receiveUpdateActiveOrders(QString pair,    QObject *sender, int SenderID);
     void receiveUpdateOrderInfo   (quint64 orderID, QObject *sender, int SenderID);
 
     void UpdateMarketTickerReply (QNetworkReply *reply);
@@ -62,9 +62,8 @@ class Exchange_wex : public Exchange
 
   signals:
     void sendTicker     (Ticker ticker);
-    void sendOrderID    (int orderID);
+    void sendOrderID    (quint64 orderID);
     void sendOrderStatus(int status);
-
   };
 
 #endif // EXCHANGE_WEX_H
